Add Makefile To Build With Oct 2020 Emscripten
The previous build instructions would not work under the current
Emscripten.  They have moved away from offering LLVM BitCode as
a target (at least, there's no obvious way to get it).

If the `-c` option is used, then the generated .wasm file will be
relocatable and can act as an object file would.  However, this
does not allow the combining of multiple source files into a single
bitcode file in the way the instructions previously had.

Also: the -lGLU option does not seem to work in this version.  The
removal of the switch resulted in a successful link and run of the
built product.

This commit is an offering of the makefile which I used to build.
Doing better would involve putting the intermediate .wasm files in
their own directories etc.  But this was just to get it built for
a quick test, so improvements are left as an exercise to the
reader.  :-)